Rename compiler tests to CLI tests
