+++ test/cli/projects/baseurl/test_src/test_lib/nested/lib_file.ts
