+++ test/cli/projects/baseurl/tsconfig.json
