+++ test/cli/projects/basic/test_src/main.ts
