+++ test/cli/projects/basic/test_src/test_lib/file.ts
