+++ test/cli/projects/basic/tsconfig.bothDirOptions.json
