+++ test/cli/projects/basic/tsconfig.json
