+++ test/cli/projects/basic/tsconfig.outDir.json
