+++ test/cli/projects/basic/tsconfig.rootDir.json
