+++ test/cli/projects/watchmode/tsconfig.json
