+++ test/cli/projects/watchmode/watch.ts
