+++ test/cli/run.ts
