+++ test/cli/testfiles/invalid_syntax.ts
