+++ test/cli/testfiles/watch.ts
