+++ test/tsconfig.json
@@ -9,5 +9,5 @@
         "noEmit": true,
         "module": "commonjs"
     },
-    "exclude": ["translation/transformation", "compiler/projects", "compiler/testfiles"]
+    "exclude": ["translation/transformation", "cli/projects", "cli/testfiles"]
 }
